+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.2 (unreleased)
 * add drag and drop support
 * improve progress indicator when loading large files
+* fix id mismatch in recent files list
 
 ## 1.2.1 (31.07.2026)
 * fix unit detection with nested parentheses, e.g. `Time (s (seconds))`